Move segment-related data structures into `segments` module
