+++ crates/pallet-subspace/src/mock.rs
@@ -38,12 +38,12 @@ use subspace_archiving::archiver::{Archiver, NewArchivedSegment};
 use subspace_core_primitives::crypto::kzg::{embedded_kzg_settings, Kzg};
 use subspace_core_primitives::pieces::{Piece, PieceOffset, Record};
 use subspace_core_primitives::segments::{
-    ArchivedHistorySegment, HistorySize, RecordedHistorySegment, SegmentCommitment, SegmentIndex,
+    ArchivedBlockProgress, ArchivedHistorySegment, HistorySize, LastArchivedBlock,
+    RecordedHistorySegment, SegmentCommitment, SegmentHeader, SegmentIndex,
 };
 use subspace_core_primitives::{
-    ArchivedBlockProgress, Blake3Hash, BlockNumber, LastArchivedBlock, PosSeed, PotOutput,
-    PublicKey, RewardSignature, SectorId, SegmentHeader, SlotNumber, Solution, SolutionRange,
-    REWARD_SIGNING_CONTEXT,
+    Blake3Hash, BlockNumber, PosSeed, PotOutput, PublicKey, RewardSignature, SectorId, SlotNumber,
+    Solution, SolutionRange, REWARD_SIGNING_CONTEXT,
 };
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::auditing::audit_sector_sync;
