+++ crates/sp-consensus-subspace/src/lib.rs
@@ -42,10 +42,12 @@ use sp_runtime_interface::{pass_by, runtime_interface};
 use sp_std::num::NonZeroU32;
 #[cfg(feature = "std")]
 use subspace_core_primitives::crypto::kzg::Kzg;
-use subspace_core_primitives::segments::{HistorySize, SegmentCommitment, SegmentIndex};
+use subspace_core_primitives::segments::{
+    HistorySize, SegmentCommitment, SegmentHeader, SegmentIndex,
+};
 use subspace_core_primitives::{
     Blake3Hash, BlockHash, BlockNumber, PotCheckpoints, PotOutput, PotSeed, PublicKey,
-    RewardSignature, SegmentHeader, SlotNumber, Solution, SolutionRange,
+    RewardSignature, SlotNumber, Solution, SolutionRange,
 };
 #[cfg(feature = "std")]
 use subspace_proof_of_space::chia::ChiaTable;
