+++ crates/subspace-fake-runtime-api/src/lib.rs
@@ -40,8 +40,10 @@ use sp_version::RuntimeVersion;
 use std::collections::btree_map::BTreeMap;
 use std::collections::btree_set::BTreeSet;
 use subspace_core_primitives::objects::BlockObjectMapping;
-use subspace_core_primitives::segments::{HistorySize, SegmentCommitment, SegmentIndex};
-use subspace_core_primitives::{PublicKey, Randomness, SegmentHeader, U256};
+use subspace_core_primitives::segments::{
+    HistorySize, SegmentCommitment, SegmentHeader, SegmentIndex,
+};
+use subspace_core_primitives::{PublicKey, Randomness, U256};
 use subspace_runtime_primitives::opaque::Block;
 use subspace_runtime_primitives::{AccountId, Balance, BlockNumber, Moment, Nonce};
