+++ crates/subspace-farmer/src/farmer_cache.rs
@@ -26,8 +26,7 @@ use std::sync::Arc;
 use std::time::Duration;
 use std::{fmt, mem};
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
-use subspace_core_primitives::segments::SegmentIndex;
-use subspace_core_primitives::SegmentHeader;
+use subspace_core_primitives::segments::{SegmentHeader, SegmentIndex};
 use subspace_farmer_components::PieceGetter;
 use subspace_networking::libp2p::kad::{ProviderRecord, RecordKey};
 use subspace_networking::libp2p::PeerId;
