+++ crates/subspace-farmer/src/farmer_cache/tests.rs
@@ -19,8 +19,9 @@ use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
-use subspace_core_primitives::segments::{HistorySize, SegmentIndex};
-use subspace_core_primitives::{LastArchivedBlock, SegmentHeader};
+use subspace_core_primitives::segments::{
+    HistorySize, LastArchivedBlock, SegmentHeader, SegmentIndex,
+};
 use subspace_farmer_components::{FarmerProtocolInfo, PieceGetter};
 use subspace_networking::libp2p::identity;
 use subspace_networking::libp2p::kad::RecordKey;
