+++ crates/subspace-farmer/src/node_client.rs
@@ -16,8 +16,7 @@ use futures::Stream;
 use std::fmt;
 use std::pin::Pin;
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
-use subspace_core_primitives::segments::SegmentIndex;
-use subspace_core_primitives::SegmentHeader;
+use subspace_core_primitives::segments::{SegmentHeader, SegmentIndex};
 use subspace_rpc_primitives::{
     FarmerAppInfo, RewardSignatureResponse, RewardSigningInfo, SlotInfo, SolutionResponse,
 };
