+++ crates/subspace-farmer/src/node_client/rpc_node_client.rs
@@ -10,8 +10,7 @@ use jsonrpsee::ws_client::{WsClient, WsClientBuilder};
 use std::pin::Pin;
 use std::sync::Arc;
 use subspace_core_primitives::pieces::{Piece, PieceIndex};
-use subspace_core_primitives::segments::SegmentIndex;
-use subspace_core_primitives::SegmentHeader;
+use subspace_core_primitives::segments::{SegmentHeader, SegmentIndex};
 use subspace_rpc_primitives::{
     FarmerAppInfo, RewardSignatureResponse, RewardSigningInfo, SlotInfo, SolutionResponse,
 };
