+++ crates/subspace-networking/src/protocols/request_response/handlers/segment_header.rs
@@ -5,8 +5,7 @@
 
 use super::generic_request_handler::{GenericRequest, GenericRequestHandler};
 use parity_scale_codec::{Decode, Encode};
-use subspace_core_primitives::segments::SegmentIndex;
-use subspace_core_primitives::SegmentHeader;
+use subspace_core_primitives::segments::{SegmentHeader, SegmentIndex};
 
 /// Segment header by segment indexes protocol request.
 #[derive(Debug, Clone, Eq, PartialEq, Encode, Decode)]
