+++ crates/subspace-service/src/sync_from_dsn/segment_header_downloader.rs
@@ -3,8 +3,7 @@ use futures::StreamExt;
 use std::collections::{BTreeSet, HashMap};
 use std::error::Error;
 use std::pin::pin;
-use subspace_core_primitives::segments::SegmentIndex;
-use subspace_core_primitives::SegmentHeader;
+use subspace_core_primitives::segments::{SegmentHeader, SegmentIndex};
 use subspace_networking::libp2p::PeerId;
 use subspace_networking::{Node, SegmentHeaderRequest, SegmentHeaderResponse};
 use tracing::{debug, error, trace, warn};
