+++ test/subspace-test-runtime/src/lib.rs
@@ -92,10 +92,10 @@ use sp_version::RuntimeVersion;
 use static_assertions::const_assert;
 use subspace_core_primitives::objects::{BlockObject, BlockObjectMapping};
 use subspace_core_primitives::pieces::Piece;
-use subspace_core_primitives::segments::{HistorySize, SegmentCommitment, SegmentIndex};
-use subspace_core_primitives::{
-    crypto, PublicKey, Randomness, SegmentHeader, SlotNumber, SolutionRange, U256,
+use subspace_core_primitives::segments::{
+    HistorySize, SegmentCommitment, SegmentHeader, SegmentIndex,
 };
+use subspace_core_primitives::{crypto, PublicKey, Randomness, SlotNumber, SolutionRange, U256};
 use subspace_runtime_primitives::{
     AccountId, Balance, BlockNumber, FindBlockRewardAddress, Hash, HoldIdentifier, Moment, Nonce,
     Signature, MIN_REPLICATION_FACTOR,
